+++ src/app/_theme/index.ts
@@ -1,1 +1,3 @@
-export { Theme } from './theme'
+export { lightTheme, darkTheme } from './theme'
+export { lightMode } from './lightMode'
+export { darkMode } from './darkMode'
